requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.19.1
openapi-codec==1.3.2
pandas==1.0.5
psycopg2==2.8.5
pyaml==20.4.0
pycodestyle==2.6.0
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.1.0
wrapt==1.12.1
